refactor(backend): share route error helper across HTTP modules
