+++ packages/backend/src/http/chatRoutes.ts
@@ -8,37 +8,17 @@
  */
 import type { IncomingMessage, ServerResponse } from 'node:http';
 import express from 'express';
+import { respondWithRouteError, type LogRequest } from './routeError.js';
 
 type RequestHandler = (
     req: IncomingMessage,
     res: ServerResponse
 ) => Promise<void>;
 
-type LogRequest = (
-    req: IncomingMessage,
-    res: ServerResponse,
-    extra?: string
-) => void;
-
 type RegisterChatRoutesDeps = {
     app: express.Express;
     handleChatRequest: RequestHandler;
     logRequest: LogRequest;
-};
-
-const respondWithRouteError = (
-    req: IncomingMessage,
-    res: ServerResponse,
-    logRequestWithContext: LogRequest,
-    error: unknown
-): void => {
-    res.statusCode = 500;
-    res.end('Internal Server Error');
-    logRequestWithContext(
-        req,
-        res,
-        error instanceof Error ? error.message : 'unknown error'
-    );
 };
 
 /**
